+++ src/tools/chronometer/chronometer.vue
@@ -4,8 +4,8 @@
       <div class="duration">{{ formatMs(counter) }}</div>
     </c-card>
     <n-space justify="center" mt-5>
-      <c-button v-if="!isRunning" secondary type="primary" @click="resume">Start</c-button>
-      <c-button v-else secondary type="warning" @click="pause">Stop</c-button>
+      <c-button v-if="!isRunning" type="primary" @click="resume">Start</c-button>
+      <c-button v-else type="warning" @click="pause">Stop</c-button>
 
       <c-button @click="counter = 0">Reset</c-button>
     </n-space>