+++ src/ui/c-input-text/c-input-text.theme.ts
@@ -1,0 +1,20 @@
+import { defineThemes } from '../theme/theme.models';
+
+export const { useTheme } = defineThemes({
+  dark: {
+    backgroundColor: '#333333',
+    borderColor: '#333333',
+
+    focus: {
+      backgroundColor: '#1ea54c1a',
+    },
+  },
+  light: {
+    backgroundColor: '#ffffff',
+    borderColor: '#e0e0e69e',
+
+    focus: {
+      backgroundColor: '#ffffff',
+    },
+  },
+});